+++ lib/garb.rb
@@ -3,51 +3,62 @@ require 'net/https'
 
 require 'cgi'
 require 'ostruct'
-require 'crack'
 
-begin
-  require 'active_support/inflector'
-  require 'active_support/deprecation'
+begin 
+  require 'yajl/json_gem' # JSON.parse
 rescue LoadError
-  require 'active_support'
+  require 'json'
 end
 
-require 'garb/version'
-require 'garb/authentication_request'
-require 'garb/data_request'
-require 'garb/account_feed_request'
-require 'garb/session'
-require 'garb/profile_reports'
-require 'garb/step'
-require 'garb/destination'
-require 'garb/goal'
-require 'garb/profile'
-require 'garb/account'
-require 'garb/filter_parameters'
-require 'garb/report_parameter'
-require 'garb/report_response'
-require 'garb/resource'
-require 'garb/report'
+module Garb
+  autoload :Destination,      'garb/destination'
+  autoload :FilterParameters, 'garb/filter_parameters'
+  autoload :Model,            'garb/model'
+  autoload :ProfileReports,   'garb/profile_reports'
+  autoload :ReportParameter,  'garb/report_parameter'
+  autoload :ReportResponse,   'garb/report_response'
+  autoload :ResultSet,        'garb/result_set'
+  autoload :Session,          'garb/session'
+  autoload :Step,             'garb/step'
+  autoload :Version,          'garb/version'
 
-require 'garb/model'
+  module Management
+    autoload :Account,     'garb/management/account'
+    autoload :Feed,        'garb/management/feed'
+    autoload :Goal,        'garb/management/goal'
+    autoload :Profile,     'garb/management/profile'
+    autoload :Segment,     'garb/management/segment'
+    autoload :WebProperty, 'garb/management/web_property'
+  end
 
-# management
-require 'garb/management/feed'
-require 'garb/management/account'
-require 'garb/management/web_property'
-require 'garb/management/profile'
+  module Request
+    autoload :Authentication, "garb/request/authentication"
+    autoload :Data,           'garb/request/data'
+  end
+end
 
-require 'support'
+# require 'garb/account_feed_request'
+# require 'garb/resource'
+# require 'garb/report'
 
 module Garb
   GA = "http://schemas.google.com/analytics/2008"
 
   extend self
 
+  class << self
+    attr_accessor :proxy_address, :proxy_port, :proxy_user, :proxy_password
+    attr_writer   :read_timeout
+  end
+
+  def read_timeout
+    @read_timeout || 60
+  end
+
   def to_google_analytics(thing)
     return thing.to_google_analytics if thing.respond_to?(:to_google_analytics)
 
-    "ga:#{thing.to_s.camelize(:lower)}"
+    "#{$1}ga:#{$2}" if "#{thing.to_s.camelize(:lower)}" =~ /^(-)?(.*)$/
   end
   alias :to_ga :to_google_analytics
 
@@ -57,13 +68,22 @@ module Garb
   alias :from_ga :from_google_analytics
 
   def parse_properties(entry)
-    entry['dxp:property'].inject({}) do |hash, p|
-      hash[Garb.from_ga(p['name'])] = p['value']
-      hash
-    end
+    Hash[entry['dxp$property'].map {|p| [Garb.from_ga(p['name']),p['value']]}]
   end
 
   def parse_link(entry, rel)
     entry['link'].detect {|link| link["rel"] == rel}['href']
   end
+
+  def symbol_operator_slugs
+    [:eql, :not_eql, :gt, :gte, :lt, :lte, :desc, :descending, :matches,
+      :does_not_match, :contains, :does_not_contain, :substring, :not_substring]
+  end
+
+  # new(address, port = nil, p_addr = nil, p_port = nil, p_user = nil, p_pass = nil)
+
+  # opts => open_timeout, read_timeout, ssl_timeout
+  # probably just support open_timeout
 end
+
+require 'garb/support'