+++ lib/garb/core_ext/array.rb
@@ -1,0 +1,15 @@
+unless Object.const_defined?("ActiveSupport")
+  class Array
+    def self.wrap(object)
+      if object.nil?
+        []
+      elsif object.respond_to?(:to_ary)
+        object.to_ary
+      else
+        [object]
+      end
+    end
+  end
+else
+  puts "ActiveSupport is suddenly defined!"
+end