+++ lib/garb/session.rb
@@ -6,7 +6,7 @@ module Garb
       # use only for single user authentication
       def login(email, password, opts={})
         self.email = email
-        auth_request = AuthenticationRequest.new(email, password, opts)
+        auth_request = Request::Authentication.new(email, password, opts)
         self.auth_token = auth_request.auth_token(opts)
       end