+++ test/unit/garb/account_feed_request_test.rb
@@ -1,9 +1,0 @@
-require 'test_helper'
-
-module Garb
-  class AccountFeedRequestTest < MiniTest::Unit::TestCase
-    context "An AccountFeedRequest" do
-      
-    end
-  end
-end