+++ test/unit/garb/session_test.rb
@@ -7,7 +7,7 @@ module Garb
       
       should "be able retrieve an auth_token for a user" do
         auth_request = mock {|m| m.expects(:auth_token).with({}).returns('toke') }
-        AuthenticationRequest.expects(:new).with('email', 'password', {}).returns(auth_request)
+        Request::Authentication.expects(:new).with('email', 'password', {}).returns(auth_request)
 
         Session.login('email', 'password')
         assert_equal 'toke', Session.auth_token
@@ -16,14 +16,14 @@ module Garb
       should "be able retrieve an auth_token for a user with secure ssl" do
         opts = {:secure => true, :account_type => 'GOOGLE'}
         auth_request = mock {|m| m.expects(:auth_token).with(opts).returns('toke') }
-        AuthenticationRequest.expects(:new).with('email', 'password', opts).returns(auth_request)
+        Request::Authentication.expects(:new).with('email', 'password', opts).returns(auth_request)
 
         Session.login('email', 'password', opts)
         assert_equal 'toke', Session.auth_token
       end
       
       should "retain the email address for this session" do
-        AuthenticationRequest.stubs(:new).returns(stub(:auth_token => 'toke'))
+        Request::Authentication.stubs(:new).returns(stub(:auth_token => 'toke'))
         
         Session.login('email', 'password')
         assert_equal 'email', Session.email
@@ -47,7 +47,7 @@ module Garb
       
       should "be able retrieve an auth_token for a user" do
         auth_request = mock {|m| m.expects(:auth_token).with({}).returns('toke') }
-        AuthenticationRequest.expects(:new).with('email', 'password', {}).returns(auth_request)
+        Request::Authentication.expects(:new).with('email', 'password', {}).returns(auth_request)
 
         @session.login('email', 'password')
         assert_equal 'toke', @session.auth_token
@@ -56,14 +56,14 @@ module Garb
       should "be able retrieve an auth_token for a user with secure ssl" do
         opts = {:secure => true, :account_type => 'GOOGLE'}
         auth_request = mock {|m| m.expects(:auth_token).with(opts).returns('toke') }
-        AuthenticationRequest.expects(:new).with('email', 'password', opts).returns(auth_request)
+        Request::Authentication.expects(:new).with('email', 'password', opts).returns(auth_request)
 
         @session.login('email', 'password', opts)
         assert_equal 'toke', @session.auth_token
       end
       
       should "retain the email address for this session" do
-        AuthenticationRequest.stubs(:new).returns(stub(:auth_token => 'toke'))
+        Request::Authentication.stubs(:new).returns(stub(:auth_token => 'toke'))
         
         @session.login('email', 'password')
         assert_equal 'email', @session.email